Check Course Explorer for classes not in dataset
I updated the bot so that if the class requested isn't found within the CSV dataset, it will check UIUC's Course Explorer page. First, it checks the page which lists the terms in which a class is offered to ensure it can reach the page. If it can, it will then fetch the most recently offered term, and grab relevant course information for the class. (Unfortunately, this includes everything except GPA).